✔ Fix tests to use the new filenames
OCR-D seems to create the files in the output file group with a new
naming scheme (based on file group name and page id); Adjust the tests
accordingly.

It's probably not the best approach to expect certain filenames, and
it would be better to use METS/the OCR-D API to ask for a certain page
and it's file, but for now we use this.